Rename range -> ranges to avoid conflicts
